requirements.txt
Python==3.7.3
pandas==0.24.2
matplotlib==3.1.2
seaborn==0.9.0
request==2.23.0
sqlalchemy==1.3.16
beautifulsoup4==4.9.1
configparser==5.0